Moving service context to servicedeployer package
